shift btc withdrawl option to settings
